requirements.txt
Flask==3.0.0
Flask-Limiter==3.5.0
Werkzeug==3.0.1
torch==2.1.2
torchvision==0.16.2
transformers==4.36.2
Pillow==10.1.0
numpy==1.26.2
gunicorn==21.2.0
